chore(frontend): rename blog service and use runtime config
