+++ frontend/server/api/blog/articles/[id].ts
@@ -1,4 +1,4 @@
-import { blogService } from '~/services/blog.services'
+import { blogService } from '~/services/blog.service'
 import type { BlogArticleData } from '../types/blog.models'
 
 /**
