+++ frontend/services/index.ts
@@ -1,2 +1,2 @@
 // Export all services
-export { blogService } from '~/services/blog.services'
+export { blogService } from '~/services/blog.service'
